requirements.txt
numpy
scipy
torch
torchvision
matplotlib
h5py
pillow
aifc
six
scikit-learn
